@@ -31,16 +31,22 @@ setup(
         'Programming Language :: Python :: 3',
         'Programming Language :: Python :: 3.6',
         'Programming Language :: Python :: 3.7',
+        'Programming Language :: Python :: 3.8',
+        'Programming Language :: Python :: 3.9',
+        'Programming Language :: Python :: 3.10',
     ],
     keywords='HEPData physics OpenData',
     packages=find_packages(exclude=['contrib', 'docs', 'tests']),
     zip_safe=False,
-    python_requires='>=2.7, !=3.0.*, !=3.1.*, !=3.2.*, !=3.3.*, !=3.4.*, !=3.5.*',
+    python_requires='>=3.6',
     install_requires=DEPS,
     setup_requires=['pytest-runner', 'pytest-cov'],
-    tests_require=['pytest', 'papermill', 'six',
-                   'pylint==2.9.6; python_version>="3"',
-                  ],
+    tests_require=[
+        'pytest',
+        'papermill',
+        'six',
+        'pylint==2.9.6',
+    ],
     project_urls={
         'Documentation': 'https://hepdata-lib.readthedocs.io',
         'Bug Reports': 'https://github.com/HEPData/hepdata_lib/issues',
